bugfix: nginx slows down in request processing for large image.
This problem was appeared in 4cbecc2.
In every request processing, the image manipulation such as
sharp and unsharp and blur ran.

The checks for the parameters below were invalid.

 * sharpen
 * unsharp
 * blur

In the earlier version than v0.7.1, the macros below return
NULL by default.

 * NGX_HTTP_SMALL_LIGHT_PARAM_GET_LIT(&ctx->hash, "sharpen")
 * NGX_HTTP_SMALL_LIGHT_PARAM_GET_LIT(&ctx->hash, "unsharp")
 * NGX_HTTP_SMALL_LIGHT_PARAM_GET_LIT(&ctx->hash, "blur")

But these retrun values are changed for other bugfix in 4cbecc2.

refs #41